Reorganise files into subdirectories.
