+++ examples/environments/Byggfile2.py
@@ -1,5 +1,5 @@
-from bygg.action import ActionContext, action
-from bygg.common_types import CommandStatus
+from bygg.core.action import ActionContext, action
+from bygg.core.common_types import CommandStatus
 import rich
