+++ examples/only_python/Byggfile.py
@@ -1,5 +1,5 @@
-from bygg.action import Action, ActionContext, action
-from bygg.common_types import CommandStatus
+from bygg.core.action import Action, ActionContext, action
+from bygg.core.common_types import CommandStatus
 
 
 def hello(ctx: ActionContext):
