+++ examples/trivial/more_things/MoreActions.py
@@ -1,5 +1,5 @@
-from bygg.action import ActionContext, action
-from bygg.common_types import CommandStatus
+from bygg.core.action import ActionContext, action
+from bygg.core.common_types import CommandStatus
 
 
 @action("testfile 2", message="testfile 2", outputs=["output2.txt"])
