+++ src/bygg/cmd/apply_configuration.py
@@ -6,10 +6,10 @@ import sys
 import textwrap
 from typing import List
 
-from bygg.action import Action
-from bygg.configuration import PYTHON_INPUTFILE, ByggFile, Environment
-from bygg.digest import calculate_string_digest
-from bygg.output import output_error, output_info, output_plain
+from bygg.cmd.configuration import PYTHON_INPUTFILE, ByggFile, Environment
+from bygg.core.action import Action
+from bygg.core.digest import calculate_string_digest
+from bygg.output.output import output_error, output_info, output_plain
 from bygg.util import create_shell_command
