+++ src/bygg/core/common_types.py
