+++ src/bygg/core/dag.py
@@ -2,33 +2,28 @@ from abc import ABCMeta, abstractmethod
 from collections import deque
 from typing import Any, Dict, Iterable, List
 
-from bygg.action import Action
+from bygg.core.action import Action
 
 
 class Dag(metaclass=ABCMeta):
     def __len__(self) -> int:
         return 0
 
-    def clear(self):
-        ...
+    def clear(self): ...
 
-    def remove_node(self, node: str):
-        ...
+    def remove_node(self, node: str): ...
 
-    def build_action_graph(self, build_actions: Dict[str, Action], action: Action):
-        ...
+    def build_action_graph(self, build_actions: Dict[str, Action], action: Action): ...
 
     @abstractmethod
     def get_ready_jobs(
         self,
         finished_jobs: Dict[str, Any],
         running_jobs: Dict[str, Any],
-    ) -> List[str]:
-        ...
+    ) -> List[str]: ...
 
     @abstractmethod
-    def get_all_jobs(self) -> Iterable[str]:
-        ...
+    def get_all_jobs(self) -> Iterable[str]: ...
 
 
 class GtDag(Dag):
