+++ src/bygg/core/digest.py
