+++ src/bygg/core/runner.py
@@ -2,14 +2,13 @@ import os
 import signal
 from typing import Callable, List
 
+from bygg.core.action import Action, CommandStatus
+from bygg.core.common_types import JobStatus
+from bygg.core.scheduler import Job, Scheduler
+from bygg.output.output import TerminalStyle as TS
+from bygg.output.status_display import on_check_failed
 import multiprocess.managers  # type: ignore
 from multiprocess.pool import ApplyResult, Pool  # type: ignore
-
-from bygg.action import Action, CommandStatus
-from bygg.common_types import JobStatus
-from bygg.output import TerminalStyle as TS
-from bygg.scheduler import Job, Scheduler
-from bygg.status_display import on_check_failed
 
 JobStatusListener = Callable[[str, JobStatus, Action, CommandStatus | None], None]
 RunnerStatusListener = Callable[[str], None]
