+++ tests/test_utils.py
@@ -1,4 +1,4 @@
-from bygg.action import Action
+from bygg.core.action import Action
 from bygg.util import create_shell_command, filenames_from_pattern
 import pytest
